refactor(Component): roll back addEvent logic
